01_etl/pyproject.toml
[tool.poetry]
name = "etltoes"
version = "0.1.0"
description = ""
authors = ["Maxim B. <wedontask@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv="0.19.2"
wheel="0.37.1"
elasticsearch = "8.1.1"
psycopg2 = "2.9.3"
pydantic = "1.9.0"
redis = "4.2.2"


[tool.poetry.dev-dependencies]
flake8="4.0.1"
black = "^22.3.0"
wemake-python-styleguide = "^0.16.1"
redis = "^4.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
